cogl_texture_get_data() copies using wrong width size.
For sliced 2D textures, _cogl_texture_2d_sliced_get_data() uses the
bitmap width as the number of bytes to copy from an intermediate
bitmap when a conversion is needed. This doesn't take into account the
bpp so it wasn't copying enough data.

Based on a patch by Brian Tarricone.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2125