cogl-material: Fix comparing the color in set_layer_combine_constant
In cogl_material_set_layer_combine_constant it was comparing whether
the new color is the same as the old color using a memcmp on the
constant_color parameter. However the combine constant is stored in
the layer data as an array of four floats but the passed in color is a
CoglColor (which is currently an array of four guint8s). This was
causing valgrind errors and presumably also the check for setting the
same color twice would always fail.

This patch makes it do the conversion to a float array upfront before
the comparison.

(cherry picked from commit 5740a5a38a4d0a0b22e62ea0035b8e360c7a6cc0)

Conflicts:

	clutter/cogl/cogl/cogl-pipeline.c