box: Implement the correct paint volume
The allocation of the ClutterBox is not enough to be used as the paint
volume, because children might decide to paint outside that area.

Instead, we should use the allocation if the Box has a background color
and then do what Group does, and union all the paint volumes of the
children.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2600
(cherry picked from commit 49d38b1e21eb868f64f213d9d5b4c6fbedd63bb2)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>