experimental clipped redraws without extension
This implements an experimental approach to supporting clipped redraws
without the aid of an extension. Instead it depends on making an
assumption about the number of back buffers provided by the driver.

To tell Clutter about the number of back buffers set
CLUTTER_SWAP_CHAIN_LENGTH=3 for example in your environment before
running a Clutter app.

Note: This is unlikely to be big win if your driver doesn't actually
support flipping back buffers and instead uses a blit to present.

Compared to the blit based clipped redraw path we have when using GLX
one significant issue is that the whole window will effectively be
damaged when using glXSwapBuffers which means the whole window could
end up being blitted if the window isn't fullscreen. Even when
compositing, because we don't have a way to report a limited damage
region the compositor will be told that the whole window has been
damage.