Support clipped redraws with flipping
This adds support for doing incremental, clipped redraws but also using
flipping to present instead of blitting.

The approach builds on Cogl's support for cogl_onscreen_start_frame(),
cogl_onscreen_get_back_buffer_age() and
cogl_onscreen_swap_buffers_with_damage().

At the start of each frame we determine how old the back buffer contents
are and if it's a reused buffer we aim to repair the old buffer based on
the history of regions that have changed since the old buffer was
created. When we present we present the back buffer we use
cogl_onscreen_swap_buffers_with_damage() which the driver can implement
using a flip and the damage region can also be passed to a system
compositor to ensure it only recomposes the region that changed.