stage-window: make it possible to damage the back buffer
This allows us to report to the backend that the stage's back buffer has
been trashed while handling picking. If the backend is keeping track of
the contents of back buffers so it can minimize how much of the stage is
redrawn then it needs to know when we do pick renders so it can
invalidate the back buffer.