stage: add dedicated pixel pick buffer
This gives each stage a dedicated 1 pixel framebuffer for handling
picking when we are only reading back a single point under the cursor.

Having a dedicated pick buffer means we can avoid damaging the contents
of the back buffer used for painting so we can re-use the contents of
old back buffers to reduce how much is drawn each frame.

Note that we don't create a full stage size pick buffer because we want
to avoid the large allocations of memory that would imply.

Note when a clutter scene is static then we will still use the back
buffer for picking so that we can do a full stage size pick render
without requiring the allocation of a dedicated buffer.