cogl/configure.ac: Adds --enable-stub-winsys option
We want to be able to split Cogl out as a standalone project but there
are still some window systems that are natively supported by Cogl. This
allows Clutter to support those window systems directly but still work
with a standalone Cogl library.

This also ensure we set the SUPPORT_STUB conditional in clutter's
configure.ac when building for win32/osx and wayland.