VS 2008 Project files for compiling Clutter
These are the VS2008 solution, property sheet and project
files for compiling Clutter.

Note that this is done in a similar way to what is now done
on GLib, GTK+, GDK-Pixbuf, ATK-so projects to compile COGL,
COGL-Pango, the main Clutter DLL, test-conformance,
test-interactive are in templates, where they are expanded
during "make dist" to simplify maintenance as files are
added and removed from these projects quite often.