VS 2010 Project files for compiling Clutter
These are the VS2010 solution, property sheet and project/
filter files for compiling Clutter.  As with the VS 2008
files, x64 and x86 compilation are supported.

Note that this is done in a similar way to what is now done
on GLib, GTK+, GDK-Pixbuf, ATK-so projects/filters to
compile COGL, COGL-Pango, the main Clutter DLL,
test-conformance, test-interactive are in templates,
where they are expanded during "make dist" to simplify
maintenance as files are added and removed from these
projects quite often.