clutter-actor: Store the queue redraw entry pointer
_clutter_stage_queue_actor_redraw returns a pointer to the
ClutterStageQueueRedrawEntry struct which it allocates. The actor is
expected to store a pointer to this so that it doesn't need to search
the list of queued redraws next time a queue redraw is called. However
_clutter_actor_queue_redraw_with_clip wasn't storing this pointer
which meant that it thought every queue redraw was the first queue
redraw. That meant that queueing a redraw with a clip would override
any previous attempts to queue a redraw instead of trying to combine
them.

This is based on commit d95258e987178 in the master branch.