actor: Avoid queueing redraws during destruction
If an actor or the stage to which it belongs are being destroyed then
there is no point in queueing a redraw that will not be seen anyway.
Bailing out early also avoids the case in which a redraw is queued
during destruction wil cause redraw entries will be added to the Stage,
which will take references on it and cause the Stage never to be
finalized.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2652
(cherry-picked from commit 19c986c0c6e0aa08ae76ab7d992fbf70ae5e44a0)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>