text: Fix the default value of "editable"
The "editable" property is documented to default to TRUE, but is
initialized to FALSE in the _init() function.

Third party code would be affected if we changed the default to be
TRUE, so we have to change the default value in the GParamSpec.

https://bugzilla.gnome.org/show_bug.cgi?id=654726
(cherry picked from commit 4d58534fbb49be6c39017df1d4a32a9184e28f68)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>