Adds initial support for stereoscopic rendering
This adds initial support for stereoscopic rendering of Clutter scenes.
The major change here is that we've introduce the internal concept of a
ClutterCamera which encapsulates a viewport, projection and view
transform. In the future the camera will also encapsulate a reference to
the destination framebuffer, but for now it's assumed that the backend
framebuffer behind stage->priv->impl is associated with all cameras.

If clutter_stage_set_stereo_enabled (stage, TRUE) is called then Clutter
will now enable stereographic rendering, using a simple toe-in approach
of modifying the view transform for each eye to model that there is a
gap between the eyes and both eyes look towards the center of the z=0
plane (This is the plane where most 2d content for a clutter application
goes)

A notable disadvantage to the simple toe-in approach is that 2D content
on the z=0 plane will look different for each eye so straight lines
for example wont appear straight to the user. A better approach we will
implement later is to setup asymetric projections for each eye that
can model the eye gap but the frustums cross at the z=0 plane such that
both eyes would have the same view of 2D content.

Currently there are three modes of output for stereoscopic content.
 » There is anaglyph rendering (for use with filter glasses with a red
   filter for the left eye and cyan filter for the right)
 » There is a vertical split mode which splits the stage and shows the
   left eye content on the left and the right eye on the right.
 » There is a horizontal split mode which splits the stage and shows
   the left eye content on the top and right eye on the bottom.

The mode can be selected using clutter_stage_set_stereo_mode().

The mode can also be explicitly overridden using the CLUTTER_STEREO_MODE
environment variable by setting it to "default", "anaglyph",
"vertical-split" or "horizontal-split". Setting this environment
variable also implicitly forces stereo rendering to be enabled.