Update config.h.win32(.in)
Added entry for HAVE_PANGO_FT2, and leave it as not defined.

Even it is possible to have Pango Freetype/FontConfig support on Windows,
but as the addition of the requirement for PangoFT2 usage is meant for
X11 platforms, and as a move to not incur extra dependencies for Clutter
on Windows (FontConfig is IMHO a library not really designed to be used
in Windows platform).