cogl-path: Clear the stroke attribute buffer when modifying the path
Commit 1515a8d7 changed cogl-path so that it will use the vertex
attribute API to stroke the path in a similar way to how it was using
the API to fill the path. However it wasn't clearing the stroke buffer
when the path is modified so it would continue to use the unmodified
stroke.

Reviewed-by: Robert Bragg <robert@linux.intel.com>

Backported from 9082dd0d71b0 of Cogl master