Require Cogl = 1.8
Cogl's API has been changed in the 1.9 cycle; specifically the
CoglVector3 type has been removed and replaced by a float[3]. This
introduced an upper bound on the version of Cogl that we can use
when building Clutter.

To remain on the safe side, let's specify that we can only build
Clutter 1.8 with a Cogl version newer than 1.8.0, but older than
1.9.2.

Distro teams will hate us, but that's how we roll.