Add a --enable-cogl2 configure option to build against Cogl 2.0
The Cogl master branch has started removing API in preparation for a
Cogl 2.0 release which will be incompatible with the current API.
Clutter is already internally mostly using the new API via
experimental features so it can more or less already compile against
2.0. However the old Cogl API is effectively part of the Clutter API
because that is what actors use for drawing so Clutter can not
silently switch to a new Cogl API. To make testing the new Cogl API
easier it is still convenient to be able to build against Cogl 2.0 for
apps that don't care about the API break.

This adds an experimental configure option to make Clutter use the
next version of Cogl. When this is enabled CLUTTER_COGL2 will be
defined from the public Cogl headers so an application can decide to
use the newer Cogl API. Clutter internally now uses this in some
places where the new API is not available as experimental API.

This patch does not update any tests which depend on the old API but
instead just skips building them when the option is enabled.

Some features of Clutter are removed when the option is enabled. These
include ClutterShader and ClutterFixed.

The Cogl compatibility header has been expanded to define more
wrappers when CLUTTER_COGL2 is defined.