Add a material compatibility when building with --enable-cogl2
The material compatibility layer for CoglPipelines has been removed in
Cogl 2.0. This patch copies it into Clutter when --enable-cogl2 is
enabled to increase the chance that a Clutter application will work
unmodified with the Cogl 2.0 API. The API for setting a user program
is not included because there are no longer any equivalent pipeline
functions. Also any API that was using CoglMaterialLayer is not
included because the wrappers from Cogl were using internal functions
to access the layers in that case.