clutter-shader-effect: Use CoglSnippets
Instead of using the legacy CoglProgram API, ClutterShaderEffect now
creates a pipeline with a CoglSnippet. The source code from the
application is written as if it was a complete shader source in the
global scope rather than as a snippet so the code is now wrapped
between a #define main .. and a #undef main so that it won't collide
with the main function that Cogl generates. This wrapped source is
then set as the declarations section on the snippet.

Previously the effect would store the application's uniform values in
a hash table and then flush them to the program whenever the effect is
painted. This has been removed and now the values are directly stored
on the pipeline which belongs to the effect. Each instance of the
effect gets its own pipeline so there is no need to have the extra
indirection of the hash table.

The effect API has a public function to retrieve the CoglProgram. In
order to avoid breaking this function the snippet is now copied into a
CoglShader and CoglProgram on demand whenever they are requested.