Wrong position of the last child of a BoxLayout.
When the first child of a box layout container is invisible, the
position of the last displayed child is wrong.

ClutterBoxLayout::allocate() first computes the sizes needed for
each of the visible children and stores those sizes into a 'sizes'
array (the array counter is incremented on visible children only).

Later on, ClutterBoxLayout::allocate() computes the position of
each visible children, using the 'sizes' array. However, it uses the
child index to access the array, instead of a 'visible child index'.

https://bugzilla.gnome.org/show_bug.cgi?id=669291